Put MockMesos in core for now and add Scheduler.connect method
